Updated Works Page
I've changed the entire theme of the page. I also had to rework the hierarchy within the div's to give a margin effect on the side of the screen. The content in the middle of screen didn't change whatsoever, only a couple font and size changes. The hover effects where lowered on the img's and removed many classes from the works.CSS file that were very unnecessary. Going to work on the Qualifications page.